schema: export the latest json schema without documentation
So far the json profile format was in movement and exporting it from
the backend made no sense as constant fluctuations where happening.
However: Recently the changes have slowed down and we have started
to write a full documentation for the profile format.
This documment is not yet finished as we are waiting for feedback from
people with more experience in writing specifications than ourselfs.
To get some community feedback however we are releaseing the json schema
and an example profile.

Signed-off-by: Mimoja <mimoja@meticuloushome.com>